+++ webview-ui/src/components/settings/SettingsView.tsx
@@ -27,6 +27,8 @@ const SettingsView = ({ onDone }: SettingsViewProps) => {
 		setAlwaysAllowBrowser,
 		soundEnabled,
 		setSoundEnabled,
+		diffEnabled,
+		setDiffEnabled,
 		openRouterModels,
 		setAllowedCommands,
 		allowedCommands,
@@ -50,6 +52,7 @@ const SettingsView = ({ onDone }: SettingsViewProps) => {
 			vscode.postMessage({ type: "alwaysAllowBrowser", bool: alwaysAllowBrowser })
 			vscode.postMessage({ type: "allowedCommands", commands: allowedCommands ?? [] })
 			vscode.postMessage({ type: "soundEnabled", bool: soundEnabled })
+			vscode.postMessage({ type: "diffEnabled", bool: diffEnabled })
 			onDone()
 		}
 	}
@@ -289,17 +292,35 @@ const SettingsView = ({ onDone }: SettingsViewProps) => {
 				)}
 
 				<div style={{ marginBottom: 5 }}>
-					<VSCodeCheckbox checked={soundEnabled} onChange={(e: any) => setSoundEnabled(e.target.checked)}>
-						<span style={{ fontWeight: "500" }}>Enable sound effects</span>
-					</VSCodeCheckbox>
-					<p
-						style={{
-							fontSize: "12px",
-							marginTop: "5px",
-							color: "var(--vscode-descriptionForeground)",
-						}}>
-						When enabled, Cline will play sound effects for notifications and events.
-					</p>
+					<h4 style={{ fontWeight: 500, marginBottom: 10 }}>Experimental Features</h4>
+
+					<div style={{ marginBottom: 5 }}>
+						<VSCodeCheckbox checked={diffEnabled} onChange={(e: any) => setDiffEnabled(e.target.checked)}>
+							<span style={{ fontWeight: "500" }}>Enable editing through diffs (very experimental!)</span>
+						</VSCodeCheckbox>
+						<p
+							style={{
+								fontSize: "12px",
+								marginTop: "5px",
+								color: "var(--vscode-descriptionForeground)",
+							}}>
+							When enabled, Cline will be able to apply diffs to make changes to files and will automatically reject truncated full-file edits.
+						</p>
+					</div>
+
+					<div style={{ marginBottom: 5 }}>
+						<VSCodeCheckbox checked={soundEnabled} onChange={(e: any) => setSoundEnabled(e.target.checked)}>
+							<span style={{ fontWeight: "500" }}>Enable sound effects</span>
+						</VSCodeCheckbox>
+						<p
+							style={{
+								fontSize: "12px",
+								marginTop: "5px",
+								color: "var(--vscode-descriptionForeground)",
+							}}>
+							When enabled, Cline will play sound effects for notifications and events.
+						</p>
+					</div>
 				</div>
 
 				{IS_DEV && (