+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Roo Cline Changelog
 
+## [2.1.15]
+
+- Incorporate dbasclpy's [PR](https://github.com/RooVetGit/Roo-Cline/pull/54) to add support for gemini-exp-1206
+- Make it clear that diff editing is very experimental
+
+## [2.1.14]
+
+- Fix bug where diffs were not being applied correctly and try Aider's [unified diff prompt](https://github.com/Aider-AI/aider/blob/3995accd0ca71cea90ef76d516837f8c2731b9fe/aider/coders/udiff_prompts.py#L75-L105)
+- If diffs are enabled, automatically reject write_to_file commands that lead to truncated output
+
+## [2.1.13]
+
+- Fix https://github.com/RooVetGit/Roo-Cline/issues/50 where sound effects were not respecting settings
+
+## [2.1.12]
+
+- Incorporate JoziGila's [PR](https://github.com/cline/cline/pull/158) to add support for editing through diffs
+
+## [2.1.11]
+
+- Incorporate lloydchang's [PR](https://github.com/RooVetGit/Roo-Cline/pull/42) to add support for OpenRouter compression
+
 ## [2.1.10]
 
 - Incorporate HeavenOSK's [PR](https://github.com/cline/cline/pull/818) to add sound effects to Cline